Style product's grid
Improved on classes and font size for the h1 elements

Added properties to the product element that uses vh instead of %
because the latter caused problems with positioning. % reduced de image
wrapper however the space remained there, however vh worked as intended.
reducing the image-wrapper size and also remiving that from the grid
element.